pyproject.toml
[project]
name = "dspy-compounding-engineering"
version = "0.1.1"
description = "DSPy implementation of the Compounding Engineering"
readme = "README.md"
authors = [
    {name = "Strategic Automation", email = "hello@strategicautomation.co.uk"}
]
license = {text = "MIT"}
keywords = ["ai", "llm", "dspy", "compounding", "engineering", "cli", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "dspy-ai",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv",
    "openai",
    "python-frontmatter",
    "jsonschema",
    "pyyaml>=6.0",
    "filelock>=3.15.4",
    "qdrant-client>=1.7.0",
    "pygments<2.19.0",
    "fastembed>=0.2.0",
    "tiktoken>=0.7.0",
]


# Optional dependencies for full agent functionality
[project.optional-dependencies]
full = [
    "dspy-ai",
    "python-dotenv",
    "openai",
    "python-frontmatter",
    "jsonschema",
    "pyyaml>=6.0",
    "filelock>=3.15.4",
]

[project.scripts]
compounding = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["agents", "workflows", "utils", "cli.py", "config.py"]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["agents", "workflows", "utils"]
